Move files correctly
