+++ src/types/index.ts
@@ -2,5 +2,6 @@ export * from './auth';
 export * from './blacklist';
 export * from './chart';
 export * from './daily';
+export * from './plot';
 export * from './profit';
 export * from './types';
