+++ src/types/types.ts
@@ -187,20 +187,3 @@ export interface PairHistory {
   /** Data end date in as millisecond timestamp */
   data_stop_ts: number;
 }
-
-export interface IndicatorConfig {
-  color?: string;
-  type?: string;
-}
-
-export interface PlotConfig {
-  main_plot: Record<string, IndicatorConfig>;
-  subplots: Record<string, Record<string, IndicatorConfig>>;
-}
-
-export interface PlotConfigStorage {
-  [key: string]: PlotConfig;
-}
-
-// eslint-disable-next-line @typescript-eslint/camelcase
-export const EMPTY_PLOTCONFIG: PlotConfig = { main_plot: {}, subplots: {} };
